+++ packages/storybook/stories/titanium-elements.stories.js
@@ -126,7 +126,7 @@ storiesOf('UI COMPONENTS|- Titanium Super Demo -', module)
           </p>
         </titanium-card>
         <br />
-        <titanium-single-action-card ?disabled=${disable} buttonTitle="Manage your data & personalization">
+        <titanium-single-action-card ?disable-action=${disable} buttonTitle="Manage your data & personalization">
           <h1>Privacy & personalization</h1>
           <p>
             Nulla posuere lacus luctus, fringilla orci vel, laoreet felis. Maecenas imperdiet urna eu consequat lobortis. Sed at ornare nibh. Donec justo urna,
@@ -134,7 +134,7 @@ storiesOf('UI COMPONENTS|- Titanium Super Demo -', module)
           </p>
         </titanium-single-action-card>
         <br />
-        <titanium-single-action-card ?disabled=${disable} buttonTitle="Manage your data & personalization">
+        <titanium-single-action-card ?disable-action=${disable} buttonTitle="Manage your data & personalization">
           <h1>Manage payment methods</h1>
           <titanium-card-list-item ?disabled=${disable} title="Mastercard">
             <horizontal-pane style="display: flex; justify-content:space-between; flex-direction: row; fill: #5f6368">