+++ packages/titanium-card/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/LeavittSoftware/titanium-elements/compare/@leavittsoftware/titanium-card@1.1.0...@leavittsoftware/titanium-card@1.1.1) (2019-06-07)
+
+
+### Bug Fixes
+
+* Corrects text color on list-item ([6e0f5eb](https://github.com/LeavittSoftware/titanium-elements/commit/6e0f5eb))
+
+
+
+
+
+# [1.1.0](https://github.com/LeavittSoftware/titanium-elements/compare/@leavittsoftware/titanium-card@1.0.0...@leavittsoftware/titanium-card@1.1.0) (2019-06-07)
+
+
+### Features
+
+* **List item:** defaults to align-items: center; ([22e9c3c](https://github.com/LeavittSoftware/titanium-elements/commit/22e9c3c))
+
+
+
+
+
+# [1.0.0](https://github.com/LeavittSoftware/titanium-elements/compare/@leavittsoftware/titanium-card@0.2.0...@leavittsoftware/titanium-card@1.0.0) (2019-06-07)
+
+
+### Bug Fixes
+
+* Fixes broken card focus in IE11 ([c0123d9](https://github.com/LeavittSoftware/titanium-elements/commit/c0123d9))
+
+
+### Features
+
+* Defaults item to flex-direction row & spacer css ([2623df9](https://github.com/LeavittSoftware/titanium-elements/commit/2623df9))
+
+
+### BREAKING CHANGES
+
+* Changes attribute disabled to disable-action in single-action-card
+
+
+
+
+
 # [0.2.0](https://github.com/LeavittSoftware/titanium-elements/compare/@leavittsoftware/titanium-card@0.1.0...@leavittsoftware/titanium-card@0.2.0) (2019-06-04)