+++ packages/titanium-data-table/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.1.0](https://github.com/LeavittSoftware/titanium-elements/compare/@leavittsoftware/titanium-data-table@3.0.3...@leavittsoftware/titanium-data-table@3.1.0) (2019-06-10)
+
+
+### Bug Fixes
+
+* Allowed flex items to wrap. Removed margin on bottom and set -16px margin left to counter 16px left on children ([45f97a3](https://github.com/LeavittSoftware/titanium-elements/commit/45f97a3))
+* Removed padding in case no slotted children are provided ([21c0ea7](https://github.com/LeavittSoftware/titanium-elements/commit/21c0ea7))
+
+
+### Features
+
+* added new table-sub-actions slot ([68edb4d](https://github.com/LeavittSoftware/titanium-elements/commit/68edb4d))
+
+
+
+
+
 ## [3.0.3](https://github.com/LeavittSoftware/titanium-elements/compare/@leavittsoftware/titanium-data-table@3.0.2...@leavittsoftware/titanium-data-table@3.0.3) (2019-06-05)
 
 **Note:** Version bump only for package @leavittsoftware/titanium-data-table